Split the device picker out, keep the form link when popups block
