+++ src/translations/en.json
@@ -1288,6 +1288,8 @@
     "device_skip": "No specific device",
     "device_filter": "Filter devices",
     "device_capture_failed": "Couldn't capture the device config; the form opens without it",
+    "device_ready": "A pre-filled GitHub issue opened in a new tab. If it didn't, use the button below.",
+    "open_form": "Open GitHub issue",
     "back": "Back",
     "browse_builder": "Device Builder issues",
     "browse_builder_desc": "Open issues filed against Device Builder.",
