+++ src/util/bug-report-prefill.ts
@@ -1,0 +1,108 @@
+import type { ConfiguredDevice } from "../api/types/devices.js";
+import { setFittedConfigParam } from "./crash-report-budget.js";
+import { devicePlatform, ESPHOME_BUG_FORM_URL, issuePlatform } from "./crash-report.js";
+import { deviceSortKey } from "./device-sort.js";
+
+/**
+ * Prefill assembly for the report-a-bug device picker: which GitHub
+ * form each bug path opens and the pure URL/facts building, kept
+ * DOM-free so the shapes are unit-testable.
+ */
+
+/** The two bug paths that carry a device config. */
+export type DeviceTarget = "builder" | "esphome";
+
+// Field ids come from each repo's `.github/ISSUE_TEMPLATE/bug_report.yml`:
+// esphome/esphome's form has `config` + `additional`; the builder form has
+// `config` + `extra` (config added in esphome/device-builder#2482). GitHub
+// silently drops unknown params, so these must track the templates.
+export const DEVICE_TARGETS: Record<
+  DeviceTarget,
+  { href: string; factsParam: "additional" | "extra" }
+> = {
+  builder: {
+    href: "https://github.com/esphome/device-builder/issues/new?template=bug_report.yml",
+    factsParam: "extra",
+  },
+  esphome: {
+    href: ESPHOME_BUG_FORM_URL,
+    factsParam: "additional",
+  },
+};
+
+// The builder template's `config` field is required; its own description
+// tells reporters to write this when no device applies.
+export const NOT_DEVICE_SPECIFIC = "not device specific";
+
+export interface PrefillContext {
+  serverVersion: string;
+  esphomeVersion: string;
+  /** Bug-form installation dropdown value; "" when unknown. */
+  installation: string;
+}
+
+/** The fully prefilled form URL for a picked device. */
+export function buildDeviceIssueUrl(
+  target: DeviceTarget,
+  device: ConfiguredDevice,
+  /** Masked YAML; "" opens the form without a config. */
+  maskedConfig: string,
+  ctx: PrefillContext
+): URL {
+  const url = deviceTargetUrl(target, ctx, device);
+  url.searchParams.set(
+    DEVICE_TARGETS[target].factsParam,
+    deviceFacts(device, target, ctx)
+  );
+  if (maskedConfig) setFittedConfigParam(url, maskedConfig);
+  return url;
+}
+
+/** Today's URL for the no-specific-device row; the builder path fills
+ *  its required config field with the template's own sentence. */
+export function skipDeviceUrl(target: DeviceTarget, ctx: PrefillContext): URL {
+  const url = deviceTargetUrl(target, ctx);
+  if (target === "builder") url.searchParams.set("config", NOT_DEVICE_SPECIFIC);
+  return url;
+}
+
+/**
+ * The facts the form's dropdowns can't carry (GitHub only prefills
+ * input/textarea fields), rendered as a bullet list.
+ */
+export function deviceFacts(
+  device: ConfiguredDevice,
+  target: DeviceTarget,
+  ctx: PrefillContext
+): string {
+  const platform = issuePlatform(devicePlatform(device));
+  return [
+    `Device: ${deviceSortKey(device)} (${device.configuration})`,
+    device.board_id && `Board: ${device.board_id}`,
+    platform && `Platform: ${platform}`,
+    device.runtime_state.deployed_version &&
+      `ESPHome running: ${device.runtime_state.deployed_version}`,
+    target === "builder" && ctx.esphomeVersion && `ESPHome: ${ctx.esphomeVersion}`,
+    ctx.installation && `Installation: ${ctx.installation}`,
+  ]
+    .filter(Boolean)
+    .map((fact) => `- ${fact}`)
+    .join("\n");
+}
+
+// Base form URL for *target* with the version param set: the builder
+// form takes the dashboard version, esphome's the device's compiled
+// version falling back to the installed core version.
+function deviceTargetUrl(
+  target: DeviceTarget,
+  ctx: PrefillContext,
+  device?: ConfiguredDevice
+): URL {
+  const url = new URL(DEVICE_TARGETS[target].href);
+  const version =
+    target === "esphome"
+      ? device?.current_version || ctx.esphomeVersion
+      : ctx.serverVersion;
+  if (version) url.searchParams.set("version", version);
+  return url;
+}
